requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
click==8.3.1
duckdb==1.4.4
fastapi==0.128.0
h11==0.16.0
httptools==0.7.1
idna==3.11
mariadb==1.1.14
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
six==1.17.0
starlette==0.50.0
tabulate==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
